continuum-node/requirements.txt
fastapi
uvicorn[standard]
websockets
aiohttp
pyyaml
requests
slowapi
